Should capture inserted text
Problem
Selecting option from the options list. Clicking "Select All" (cmd + A) in the field and clicking any letter on the keyboard. Selection is cleared, but the first letter is not captured in the text field.

Solution
Capturing TextChanged event with entered text even if current selection exists. If developer does not clear the "selected" by themselves the label of "selected" is anyway shown.